Adding first revision of tab support (also copied over the latest from the trunk).
View contains all of the changes for this feature. The parts working correctly are: opening a new tab "tabnew", deleting a tab (segfault when removing the last one), switching between tabs with "gt" "gT" "tabn[ext]" "tabp[revious]". Existing functionality was maintained except the behavior for 'g' which was repurposed to the standardard vim keys for switching tabs.

Fixes still required: assertion failure (attempted to reparent vbox widget) when trying to open the same document in different tabs, more bug testing, customizing the look and feel.